tests/NerdStore.Vendas.Application.Tests/OrderCommandHandlerTests.cs
using FluentAssertions;
using MediatR;
using Moq;
using Moq.AutoMock;
using NerdStore.Vendas.Application.Commands;
using NerdStore.Vendas.Domain;
using NerdStore.Vendas.Domain.Interfaces;
using System;
using System.Threading;
using System.Threading.Tasks;
using Xunit;

namespace NerdStore.Vendas.Application.Tests
{
    public class OrderCommandHandlerTests
    {
        [Fact]
        public async Task AddItem_NewOrder_ShouldExecuteSuccessfully()
        {
            // Arrange
            var orderCommand = new AddOrderItemCommand(Guid.NewGuid(), Guid.NewGuid(), "Test product", 2, 50);
            var mocker = new AutoMocker();
            var handler = mocker.CreateInstance<OrderCommandHandler>();
            mocker.GetMock<IOrderRepository>().Setup(r => r.UnitOfWork.Commit()).Returns(Task.FromResult(true));

            // Act
            var result = await handler.Handle(orderCommand, CancellationToken.None);

            // Assert
            result.Should().BeTrue();
            mocker.GetMock<IOrderRepository>().Verify(r => r.Add(It.IsAny<Order>()), Times.Once);
            mocker.GetMock<IOrderRepository>().Verify(r => r.UnitOfWork.Commit(), Times.Once);
            //mocker.GetMock<IMediator>().Verify(m => m.Publish(It.IsAny<INotification>(), CancellationToken.None), Times.Once);
        }
    }
}
